models/team16_APRIL-AIGC/requirements.txt
# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0
# fastgen 0.1.0 default:
# old: diffusers==0.35.1
# old: transformers==4.49.0
# 为了运行flux2进行更新, 更新后可能原来的模型会出错
# These dependencies are installed by setup.py when calling `pip install -e .`
wandb[media]
git+https://github.com/huggingface/diffusers.git@64e2adf
transformers==5.1.0
accelerate
safetensors
scipy
einops
jupyter
hydra-core
omegaconf
imageio
tqdm
webdataset
av
loguru
boto3
timm
ftfy
opencv-python-headless
sentencepiece
nvidia-ml-py
pillow
torchvision
huggingface_hub
attrs
numpy<2.0.0
torch>=2.6.0

matplotlib